move setup_services back to media_player.py
