+++ homeassistant/components/bluesound/media_player.py
@@ -8,7 +8,7 @@ from collections.abc import Callable
 from contextlib import suppress
 from datetime import datetime, timedelta
 import logging
-from typing import Any
+from typing import TYPE_CHECKING, Any, NamedTuple
 
 from aiohttp.client_exceptions import ClientError
 from pyblu import Input, Player, Preset, Status, SyncStatus
@@ -26,6 +26,7 @@ from homeassistant.components.media_player import (
 )
 from homeassistant.config_entries import SOURCE_IMPORT
 from homeassistant.const import (
+    ATTR_ENTITY_ID,
     CONF_HOST,
     CONF_HOSTS,
     CONF_NAME,
@@ -37,6 +38,7 @@ from homeassistant.core import (
     DOMAIN as HOMEASSISTANT_DOMAIN,
     Event,
     HomeAssistant,
+    ServiceCall,
     callback,
 )
 from homeassistant.data_entry_flow import FlowResultType
@@ -48,9 +50,20 @@ from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 from homeassistant.util import Throttle
 import homeassistant.util.dt as dt_util
 
-from . import BluesoundConfigEntry
-from .const import ATTR_BLUESOUND_GROUP, ATTR_MASTER, DOMAIN, INTEGRATION_TITLE
+from .const import (
+    ATTR_BLUESOUND_GROUP,
+    ATTR_MASTER,
+    DOMAIN,
+    INTEGRATION_TITLE,
+    SERVICE_CLEAR_TIMER,
+    SERVICE_JOIN,
+    SERVICE_SET_TIMER,
+    SERVICE_UNJOIN,
+)
 from .utils import format_unique_id
+
+if TYPE_CHECKING:
+    from . import BluesoundConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -81,6 +94,29 @@ PLATFORM_SCHEMA = MEDIA_PLAYER_PLATFORM_SCHEMA.extend(
     }
 )
 
+BS_SCHEMA = vol.Schema({vol.Optional(ATTR_ENTITY_ID): cv.entity_ids})
+
+BS_JOIN_SCHEMA = BS_SCHEMA.extend({vol.Required(ATTR_MASTER): cv.entity_id})
+
+
+class ServiceMethodDetails(NamedTuple):
+    """Details for SERVICE_TO_METHOD mapping."""
+
+    method: str
+    schema: vol.Schema
+
+
+SERVICE_TO_METHOD = {
+    SERVICE_JOIN: ServiceMethodDetails(method="async_join", schema=BS_JOIN_SCHEMA),
+    SERVICE_UNJOIN: ServiceMethodDetails(method="async_unjoin", schema=BS_SCHEMA),
+    SERVICE_SET_TIMER: ServiceMethodDetails(
+        method="async_increase_timer", schema=BS_SCHEMA
+    ),
+    SERVICE_CLEAR_TIMER: ServiceMethodDetails(
+        method="async_clear_timer", schema=BS_SCHEMA
+    ),
+}
+
 
 def _add_player(
     hass: HomeAssistant,
@@ -98,12 +134,12 @@ def _add_player(
         hass.async_create_task(bluesound_player.async_init())
 
     @callback
-    def _start_polling(event: Event):
+    def _start_polling(event: Event | None = None):
         """Start polling."""
         bluesound_player.start_polling()
 
     @callback
-    def _stop_polling(event: Event):
+    def _stop_polling(event: Event | None = None):
         """Stop polling."""
         bluesound_player.stop_polling()
 
@@ -173,6 +209,33 @@ async def _async_import(hass: HomeAssistant, config: ConfigType) -> None:
             "integration_title": INTEGRATION_TITLE,
         },
     )
+
+
+def setup_services(hass: HomeAssistant) -> None:
+    """Set up services for Bluesound component."""
+
+    async def async_service_handler(service: ServiceCall) -> None:
+        """Map services to method of Bluesound devices."""
+        if not (method := SERVICE_TO_METHOD.get(service.service)):
+            return
+
+        params = {
+            key: value for key, value in service.data.items() if key != ATTR_ENTITY_ID
+        }
+        if entity_ids := service.data.get(ATTR_ENTITY_ID):
+            target_players = [
+                player for player in hass.data[DOMAIN] if player.entity_id in entity_ids
+            ]
+        else:
+            target_players = hass.data[DOMAIN]
+
+        for player in target_players:
+            await getattr(player, method.method)(**params)
+
+    for service, method in SERVICE_TO_METHOD.items():
+        hass.services.async_register(
+            DOMAIN, service, async_service_handler, schema=method.schema
+        )
 
 
 async def async_setup_entry(
